refactor: replace instances to use useCurrentUser where necessary; remove second declaration of URI
